Application: implement escape sequences
The standard specifies that the escape sequences \s, \n, \t, \r and \\
shall be recognised for the values of types string, localestring and
iconstring and that the escape sequence \; (including the other ones)
shall be recognised for values of type string(s). The previous
implementation copies the values verbatim and therefore is nonconforming
to the standard.